added checks whether the id matches the logged in user

+++ backend/Autovibe.API/Controllers/CarsController.cs
@@ -181,8 +181,19 @@ namespace Autovibe.API.Controllers
                 return NotFound();
             }
 
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
             
-            
+            if(userIdClaim == null && int.TryParse(userIdClaim.Value, out int loggedInUserId))
+            {
+                return Unauthorized("User not authenticated.");
+            }
+            else
+            {
+                if(car.UserId != loggedInUserId){
+                    return Unauthorized("You are not the owner of this car.");
+                }
+            }
+
             car.Make = updateDto.Make;
             car.Model = updateDto.Model;
             car.Year = updateDto.Year;
@@ -244,6 +255,18 @@ namespace Autovibe.API.Controllers
             {
                 return NotFound();
             }
+
+            var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier);
+            
+            if(userIdClaim == null && int.TryParse(userIdClaim.Value, out int userId))
+            {
+                return Unauthorized("User not authenticated.");
+            }
+            if(car.UserId != userId)
+            {
+                return Unauthorized("You are not the owner of this car.");
+            }
+            
             _context.Cars.Remove(car);
             await _context.SaveChangesAsync();
